Move specifiers module from codegen to core
