+++ brian2/codegen/__init__.py
@@ -1,6 +1,5 @@
 from .languages import *
 from .functions import *
-from .specifiers import *
 from .statements import *
 from .templating import *
 from .translation import *
