+++ brian2/codegen/translation.py
@@ -16,11 +16,11 @@ The input information needed:
 '''
 import re
 
+from brian2.core.specifiers import (Value, ArrayVariable, OutputVariable,
+                                    Subexpression, Index)
 from brian2.utils.stringtools import (deindent, strip_empty_lines, indent,
                                       get_identifiers)
 
-from .specifiers import (Function, Value, ArrayVariable, OutputVariable,
-                         Subexpression, Index)
 from .statements import Statement
 
 __all__ = ['translate', 'make_statements']
