+++ brian2/core/specifiers.py
